Consistently use "library" and "package" (#665)
Update usage of the words "library" and "package" to be more consistent.
The idea is that when talking about the code/API we're talking about the
"library" and when we're talking about the thing that's published to npm
it's a "package" (since it's the "Node _Package_ Manager").

The scope of this commit is the entire repository: in-code documentation
or out-of-code documentation, or anything else (e.g. the issue template)

Also sneak in a fix for a type in the Contributing Guidelines'
Compatibility Testing section ("as" -> "is").