Remove use of `utf8` from library
Replace the use of utf8 from the library by the Node.js built-in package
`util`'s `TextEncoder` [1]. This class was added in Node.js v8.3.0 so is
safe to use from a compatibility perspective.

Since the C shell specific unit test (currently) still uses the package,
it has been changed into a devDependency.

--
1. https://nodejs.org/api/util.html#class-utiltextencoder