Add C shell support (#659)
Add support for C shell [1] (on Unix systems), or `csh`, with an
implementation based primarily on fuzzing against `csh`using all
currenlty available fuzz targets in this project.

Unit and integration test fixtures were created based on 1) the fixtures
available for other Unix shells, adjusted based on the behaviour of the
`csh`-specific escaping discovered based on fuzzing, and 2) additional
cases for characters that should be escaped for `csh`; correspondingly
such cases were also added for all other Unix shells.

Additionally, by including C shell in the `test/_constants.cjs` file its
covered by other test suites such as the end-to-end tests.

One notable thing about escaping for C shell is the code that can be
found in `src/unix.js` at line 103. The preceding comment explains this
in detail but in essence, due to a bug, characters with the bytes 0xA0
in the utf-8 encoding are to be quoted individually.
As checking for this requires the Node.js built-in package `util` [2],
`util` is added as an "external" package in the rollup.js configuration
to avoid unintentinally bundling it into `index.cjs`.
There is a special unit test for this detail in a C shell specific unit
testfile that generates arbitrary strings with arbitrary 0xA0-containing
characters. Additionally, the two new fuzz corpus entries both relate to
this behaviour - one with 0xA0 in the character's -8 encoding second
position and one with 0xA0 in the character's utf-8 encoding third
positionl.

--
1. https://en.wikipedia.org/wiki/C_shell
2. https://nodejs.org/api/util.html