Update tests for arguments with exclamation marks
Update the test fixtures for `csh` to include examples of arguments with
the substring `"\!"` to ensure they are escaped properly. In particular,
the test case `"a\!"` is currently not escaped correctly by Shescape, so
this test acts as a regression test. The other tests are included for
completeness and to avoid regression against the current implementation
when fixing the current bug.

Additionally, include a new fuzz corpus entry which shows that it is
really only trailing `"\!"` substrings that aren't handled correctly. It
will blow up if `"\!"` in the middle of an argument is over-escaped.