Improve `shescapeOptions` arbitrary
Revert changes from 212767b and update the `shescapeOptions` arbitrary
to generate option objects that are more likely to be realistic by
telling the `fc.object` arbitrary to generate values using the
`unixShell` and `windowsShell` arbitraries. This makes it more likely to
have the `"shell"` option receive a real shell value and thus more
likely to quote-escape for each shell.

This is towards the same goal as 212767b, which aimed to improve
coverage of the `} else if (quoted) {` branches in integration tests.
In that commit it was achieved by explicitly testing all shells. Here,
the approach shifted to making the arbitrary used more representative.
This achieves the goal of increasing integration test coverage, as well
as improves any test using this arbitrary - bonus points for that :)