Allow imports for fuzz targets in e2e tests
Avoid lint warnings about importing from test files in the case of
importing from fuzz targets. This is more of a temporary solution
to minimize changes and because it's not trivial to find a good
alternative import structure (note that fuzz targets can't import
AVA test macros since AVA is incompatible with CJS).