Create initial testing utilities
Initiate "testing utilities" - helper functionality for testing purposes
for library users. This aims to provide users of shescape with stubbed
implementations of shescape for use in their tests.

The reasoning for this is twofold:
1. While it's not too hard to create a stub/mock for shescape, there's
   some nuance that library users might not be aware of.
2. Using shescape's main implementation in tests might not be desirable
   as it depends on the environment and system setup. Providing test
   doubles hopefully encourages library users to avoid using it in,
   mainly, unit tests.

This initial implementation contains some exact code duplication,
deduplication is left as an improvement for later.