Provide testing utilities to library users (#710)
Initiate "testing utilities" - helper functionality for testing purposes
for library users. This aims to provide users of shescape with stubbed
implementations of shescape for use in their tests.

The reasoning for this is twofold:
1. While it's not too hard to create a stub/mock for shescape, there's
   some nuance that library users might not be aware of.
2. Using shescape's main implementation in tests might not be desirable
   as it depends on the environment and system setup. Providing test
   doubles hopefully encourages library users to avoid using it in,
   mainly, unit tests.

To allow for some re-use of existing code in the testing utilities
implementation, existing code has been refactored slightly. This amounts
to move code into a new module called `reflection.js` and importing from
it in pre-existing module.

The testing utilities are made available by importing from
`"shescape/testing"`. To provide backwards compatibility with older Node
versions the testing utilities are transpiled to CommonJS (separately
from the main library's `index.cjs` file). Configurations, scripts, and
developer docs have been updated to account for this new file (named
`./testing.cjs`).

The testing utilities are documented in the `docs/testing.md` file,
following the existing documentation approach. The testing documentation
is referenced in the `README.md` for visibility.

This change is included in the `CHANGELOG.md` as it is a change that can
be used by users of the library.