Distinguish between latest and non-latest versions
If there is a vulnerability in a supported version, but that version is
not the latest version - i.e., latest isn't affected it can be disclosed
publicly.

Disclosing it public makes it so users can be aware sooner and it avoids
the hassle involved in trying to keep things secret.