Update `SECURITY.md` (#723)
* Distinguish between latest and non-latest versions

If there is a vulnerability in a supported version, but that version is
not the latest version - i.e., latest isn't affected it can be disclosed
publicly.

Disclosing it public makes it so users can be aware sooner and it avoids
the hassle involved in trying to keep things secret.

* Suggest private disclosure if it's not clear if a latest version is
  affected

* Include section on things to consider for a report

This is to give reporters an idea of what they can do to be the most
helpful to the project. The items at the beginning of the list are
pretty standard, but later items might not be immediately obvious.

* Drop "supported" column

The value can be derived from the end-of-life and today's date. As a
result, the value present in the table can only get out of date.

* Specify scope of supported versions table

This is to prevent someone looking at an old version of this file
wondering if newer version are forgotten about or not.

* Consistently use "security issue" (over "security bug" or "problem")