Update egress policy for the CodeQL job (#765)
Add `objects.githubusercontent.com:443` to the allowed endpoints of the
egress policy for the "Checks / CodeQL" job. This follows unexpected
failures in that job (i.e. without any changes) when it attempted to
(re-)download CodeQL. Since this endpoint is already allowed for various
other jobs, adding it is considered a non-issue.