Pre-format in pre-commit hook
Update the pre-commit hook to format before stashing. This has the
effect of avoiding stash-merge conflicts due to the stash pop at the end
of the hook in some cases. This is considered not harmful as,
eventually, all code (staged or not) must be formatted.

In this repository the performance impact of formatting twice is not
considered to be a big deal, if it is it would be more effective to
add some logic to only format dirty files to begin with.