Relicense tests and scripts (#789)
Change the license of all tests and scripts that used to use The
Unlicense [1] license to the MIT [2] license. Both are permissive
licenses, but the MIT license has both clearer language (from a legal
perspective) and is _more_ permissive - per the Blue Oak Council [3].

The MIT license was chosen as the replacement permissive license as it's
common in the JavaScript ecosystem and at the same time a reasonably
good permissive license according to the Blue Oak Council [3].

All of these files were only edited by me and there's therefor no other
people with a claim to the license used. Not to mention (again) that
both licenses are permissive.

--
1. https://opensource.org/license/unlicense/
2. https://opensource.org/license/mit/
3. https://blueoakcouncil.org