Set up continuous SAST with njsscan (#793)
* Set up continuous SAST with njsscan

Add a new CI job that runs the SAST tool njsscan[1]. A tool that is a:

> Static security code scanner (SAST) for Node.js applications powered
> by libsast and semgrep.

The Action is based on a tool with the same name: njsscan[2]. This is
also available as a CLI (not tried as of this commit). The tool is aimed
at scanning Node.js code specifically.

This is mainly intended as a test for the tool, as found in GitHub's
"Choose aw workflow" UI for security. Yes, this project already uses
Semgrep[3], but not libsast[4]. Per my understanding this scanner builds
on top of Semgrep with custom patterns for this reason we're testing it
in addition to Semgrep.

Note: the Action `jinabraham/njsscan-action` does not use semantic
versioning as of this commit, just v1, v2, etc.

From testing it appears the `|| true` is necessary, this came from the
example workflow.

* Configure harden-runner for the `njsscan` job

Based on observed traffic and egress policies of other, similar, jobs.

--
1. https://github.com/ajinabraham/njsscan-action
2. https://github.com/ajinabraham/nodejsscan
3. https://semgrep.dev/
4. https://github.com/ajinabraham/libsast