Set up continuous scanning with Trivy
Add a new CI job to continuously run Trivy [1] using the official Trivy
Action [2]. The job configuration is based on the Action documentation
as well as the example workflow from the GitHub Actions example
workflows UI.

Trivy provides SBOM generation, known vulnerability scanning,
infrastucture as code misconfiguration detection, secrets detection, and
license detection and compliance. Here, vulnerability scanning and
secrets detection are enabled. IaC not because this project has none.
License scanning not because it requires configuration (and this project
already uses licensee for this topic).

Note: the trivy-action uses the tag format `[0-9]+.[0-9]+.[0-9]+`,
without a leading "v".

This is mainly intended as a test for the tool.

--
1. https://github.com/aquasecurity/trivy
2. https://github.com/aquasecurity/trivy-action