Set up continuous auditing with the OSSF Scorecard Action
Add a new CI job to continuously audit the repository using the OSSF
Scorecard Action [1]. This Action performs the checks [3] of the OSSF
Scorecard project [2]. The job configuration is based on the Action
documentation as well as the example workflow from the GitHub Actions
example workflows UI.

The OSSF Scorecard checks [3] look for various "bad practices" in the
current project and reports and them.

The Branch-Protection check [3] is not enabled as it requires a PAT [3],
the risk associated with this are currently not deemed acceptable for
the benefits it could provide.

As this job lives in the `reusable-audit.yml` workflow, the `check.yml`
workflow has been updated to temporarily use the reusable workflow from
the current branch in order to test out the changes.

The audit policy is to be set following a run of the job based on
observed traffic.

This is mainly intended as a test for the tool.

--
1. https://github.com/ossf/scorecard-action
2. https://github.com/ossf/scorecard
3. https://github.com/ossf/scorecard/blob/cf0533a/docs/checks.md
4. https://github.com/ossf/scorecard-action/tree/80e868c#personal-access-token-pat-requirements-and-risks