Update ESLint configuration
Enable the in 41.1.0 of eslint-plugin-jsdoc rule `informative-docs` [1]
which aims to detect and report on documentation that's trivial/does not
add anything beyond the names in the source code. It has been enabled
with the goal of having better documentation. It has been configured
explicitly to allow for easier adjustment based on need in the future.

--
1. https://github.com/gajus/eslint-plugin-jsdoc/tree/cd01546#informative-docs