Publish with provenance from CI
Update the CI-based publishing strategy for this project to use the new
`--provenance` option - setup based on [1]. The egress policy for the
job has been updated to allow for signing (fulcio, rekor, & tuf) (and
also, preemptively, for downloading Node.js, when necessary).

The goal here is to increase linkability between the npm package and the
GitHub repository as well as the specific commit and CI workflow that
resulted in a specific version.

--
1. https://docs.npmjs.com/generating-provenance-statements