Don't upload coverage of e2e tests to Codecov
Remove end-to-end test coverage from the Codecov configuration file and
stop uploading coverage data for the end-to-end tests to Codecov. This
follows from observed errors in the Codecov dashboard, following [1] it
seems this is due to the fact that the end-to-end tests report coverage
on generated(/transpiled) files exclusively. Without access to the
source file, Codecov doesn't know what to do with the data. As such path
fixing [2] isn't going to help either.

This can(/should) be added back when e2e tests run on non-generated code

--
1. https://docs.codecov.com/docs/error-reference
2. https://docs.codecov.com/docs/fixing-paths