Update benchmarks
1. Change the pre-existing benchmark suite to more accurately reflect
real usage. That is, with the current API, get and escape every time;
with the experimental API, get once and escape every time.

2. Create a new benchmark suite aimed at comparing the current and
experimental APIs. Currently targeting Unix. Uses a hardcoded config to
compare the two APIs for a given shell and given options. This suite
also reports back which one was the fastest.