Create benchmark suite for end-user API comparison
Create a third benchmark suite. This one is targeted at comparing the
new APIs. Example results can be found below. The suspected reason
behind the significant performance difference is that the new API only
looks up the shell once, since this involves a call to (a Node.js
implementation of) `which(1)` the new API is much faster.

After changing the following line in `package.json`:

    - "benchmark": "node bench/bench.js",
    + "benchmark": "node bench/api.js",

Example run results:

    $ npm run benchmark

    > shescape@1.6.6 benchmark
    > node bench/api.js

    current x 2,634 ops/sec ±2.04% (70 runs sampled)
    experimental x 18,645,358 ops/sec ±0.10% (99 runs sampled)
    Fastest is experimental