Simplify replacements in accordance with regexp/prefer-lookaround
Update some of the replacement in Windows specific code to equivalent
replacements that preserve more of the string (replace fewer characters)

This was reported by the `regexp/prefer-lookaround` rule after some
refactoring of the related code on a separate branch. I haven't yet
figured out why this problem isn't being reported by the
esling-plugin-regexp with the current implementation.