Update unit tests for Windows logic
Mirroring the unit test changes for Unix logic related to the internal
refactoring.

Also improve the Windows benchmark to use variable naming in-line with
the rest of the project. As well as an improvement to fs path building
for Unix unit tests (following the same improvement on the Windows unit
test side).