Refactor internals in preparation for a more performant API (#870)
* Lift interpolation and quoted checks

Refactor the internal escaping/quoting functionality to left the
evaluation of the `interpolation` and `quoted` options up such that it
becomes possible to get an escape/quote function for a specific
configuration (rather than calling an escape/quote function with that
configuration)

The goal of this change is to enable a more performand API based on the
lift implemented here, namely evaluating escape configuration once up
front.

* Benchmark compare new/old escape logic

Add benchmarks using the new `getEscapeFunction` implementation to
compare to the old API against the new API.

* Create new benchmark suites aimed at comparing new/old escape logic

Create two new benchmark, one for Unix and one for Windows, aimed at
comparing the old against the new (internal) API for escaping for a
given setup (shell, options, argument) and reporting back the fastest
implementation. This is intended for comparing performance during
development. As of this commit, running these benchmarks for any shell
should (nearly) always report back that the "[f]astest is 'new'".