Provide injection strings for testing purposes
Include an exported value in the `shescape/testing` module that is a
list of example shell injection strings for use in testing to get
limited guarantee that shell injection is not possible. There's a lot
of caveats to this, not least of which is the license under which this
library is available ("Covered Software is provided under this License
on an _as is_  basis").