Update testing utilities documentation
Add a new section to the testing utilities on fixtures, currently
covering the only available fixture `injectionStrings`. Also rewrite
the document as the old structure no longer works (e.g. the "Why"
section does not apply to the fixtures, so shouldn't be a second level
heading).

Also correct the name of the exported value to be plural (as it's a
list).