Provide injection strings for testing purposes (#902)
* Provide injection strings for testing purposes

Include an exported value in the `shescape/testing` module that is a
list of example shell injection strings for use in testing to get
limited guarantee that shell injection is not possible. There's a lot
of caveats to this, not least of which is the license under which this
library is available ("Covered Software is provided under this License
on an _as is_  basis").

* Use example injection strings in the end-to-end tests

Update the end-to-end (e2e) tests to use the list of `injectionStrings`
made available through `shescape/testing` instead of it's own separate
list. Besides reducing duplication, it also ensures shescape always
actually protects against the list of strings offered to users for
testing.

* Update testing utilities documentation

Add a new section to the testing utilities on fixtures, currently
covering the only available fixture `injectionStrings`. Also rewrite
the document as the old structure no longer works (e.g. the "Why"
section does not apply to the fixtures, so shouldn't be a second level
heading).

* Update the changelog