API design for flag/option "escaping"
Update `index.{d.ts,js}` with updated documentation and types for
supporting the "escaping" of flags and options. Because it won't be
really "escaping" (i.e. "escaping" `--foobar` as `\-\-foobar` is
ineffective) the option is instead named "flagProtection", because
it protects against flag(/option) injection.

The default value for this option will be `false`. First and foremost
because `true` would border on being a breaking change from my point
of view. Secondly (and relatedly), it's not obvious that this is always
necessary. For example, per `docs/tips.md`, using the "--" option would
be preferable. That being said, defaulting to `true` should be
reconsidered in a future major version release as a "secure by default"
consideration.

Some changes to the tests are included as a way to show what expected
is. These tests are created at the integration level following the fact
that so far there's only API design (implementation is unknown, so unit
tests are not yet possible to create; end-to-end is less applicable
given their current nature, though not out of the question for the
future). Note that, lacking an update to the library implementation,
these new tests are currently expected to fail (consider this a TDD
commit :)).

Lastly, the `shescapeOptions` arbitrary  has been updated in accordance
with the changes to the API.