Lift `options.quoted` out of the internal `options` object
Refactor the internals of quoting and escaping to simplify the start to
finish process of the `quote` and `quoteAll` API functions. In
particular, this moves away from having the `quoted` option internally
for escaping and towards using the internal quoting logic to do the
escaping as well.

Note that the refactoring in this commit is incomplete and only covers
the system-specific escaping and quoting logic.