Change `getQuoteFunction` to return an escape-quote pair
Update the `getQuoteFunction` interface to return a pair of functions, 1
to escape and 1 to quote, instead of a single function. Update all
implementations and corresponding tests in accordance with this
interface change.

The reason for this can be found in the changes to `main.js`, where, if
flag protection is enabled, the process is:
1. Escape the arg
2. Apply flag protection to the arg
3. Quote the arg
Which is the only valid process.

This iterates on the changes in [1], where `escapeForQuoted` functions
were inlined into to `quoteArg` functions. Instead of reverting this,
this keeps the quoting related functionality scoped to the
`getQuoteFunction` implementation. This keeps `getEscapeFunction`
simpler and also avoids the invalid `options` value
`{interpolation:true,quoted:true}`.

--
1. abaf4be0d8bc6d01875fb7008e030e469979bc7e