Revert c316177615cbca1adbb65c5f1377d82282154f56
Revert the previous commit and instead test what happens if the
problematic escape is removed for PowerShell. This will fail the unit
tests (and thus most of the CI), but fuzzing should still run in the CI
which is what I'm most interested in here.

For reference, see [1] where the escape being removed here was
originally added.

--
1. fcba4eee83b53dee0e4060f8e84e4ea02d51efde