Switch to Shell-specific `stripFlagPrefix` functions
Switch from a platform wide `stripFlagPrefix` fn to a shell-specific one
because the implementation for PowerShell appears to have to be unique.
This is somewhat unfortunate as the protection mechanism shouldn't be
shell specific since flags are not something determined by the shell but
rather by the program, which is more of a system-level thing (or,
arguably, a global thing).

Further refactoring could be explored to see if the idea that flag
protection is system wide can be re-encoded in the source code.

Reverts the part of [1] that wasn't itself a revert of an earlier commit

--
1. 875127e29174f6d6756ffd3535ef8d782511de21