Update transitive (dev)dependencies (#934)
Try out `npm update` to update transitive dependencies, like [1]. Follow
the update with `npm dedupe` as an optimization.

Unfortunately, `npm update` does not support omitting production
dependencies (for this project: `which`). So, to keep those where they
are I manually uninstalled, re-install with `--save-exact`, and added
back the `^` range. This lead to no changes in `package.json`, after
which `npm install` synced the lockfile once more.

--
1. de3139e379c8757e6296203f786be84e9d9514e9