Add support for flags & options protection (#908)
Add support for protection against flag/option injection through a new
`flagProtection` option. The option will extend escaping and quoting
functionality with extra protection against flag/option injection. That
is, it aims to protect against an attacker, for example, providing an
attack string such as `--verbose` to gain extra information about the
target system.

It should be noted that it is considered to provide 100% flag protection
in practice. The reason for this is that what is considered a flag or
option is specific to the implementation of the target software. This is
as opposed to other escaping done by this software, which is shell
specific. The name was chosen to reflect this, it only provides
"protection" but isn't perfect. The protection only protects against
common flag/option strings on the given platform.

Flag protection is disabled by default as enabling it is considered a
breaking change.