Refactor and test new `checkedToString` implementation
Based on mutation testing results, which uncovered 4 mutants in the
previous implementation, this refactors the `checkedToString` (and
correspondingly `isStringable`) implementation and improves related test
suites.

First, the refactoring of the source code. Mutation testing showed
mutants on lines 74 and 78 of `reflection.js` replacing the return value
by `[]`. While not particularly interesting, refactoring to accommodate
these mutants has made the code more JavaScript-y (instead of Golang-y).
So I'm overall happy with this change.

Second, the new `illegalArguments` case. This followed from a mutant
that appeared after the refactoring above. In particular, on what is now
line 82 of `reflection.js` a constant `true` would work as a result of
the introduced usage of the value `null` representing a non-stringable
value. The `toString` implementation returning a number catches this
problem because it would not result in a(/the expected) type error.

Lastly, the new tests in `escape.test.js` and `quote.test.js` address
two more mutants uncovered in `reflection.js`. Namely that the new check
for strings at the start of `checkedToString` is unnecessary (can be
skipped with a constant `false` condition or by emptying the block). The
new tests address this by overriding the `toString` method of strings to
not return a string.
Beyond addressing the mutants I think this test makes sense because I
think it would be unexpected if a string is rejected because its
toString value isn't a string (which previously it would!).

Technically, this could be considered a bugfix. However, the bug is so
obscure (since it requires overriding `String.prototype.toString` to not
return a string) that I don't intent on recording it in the changelog.
Equally, this could be considered a breaking change. However, since it's
1) a bug fix, and 2) makes the library work as intended in more cases, I
definitely don't consider this a breaking change.

Ow, and this re-orders imports in `main.js` to be alphabetical.