Reward providing arguments as strings (#949)
In most cases provided arguments should be strings. So, reward providing
string with slightly better performance in the form of fewer conditions.

In particular, `checkedToString` will now return early if the provided
argument is a string _instead of_ performing a `undefined`/`null` check,
seeing if `toString()` can be called, calling `toString()`, and only
then returning the string value.

The new flow still performs all the other checks (non-`null`/`undefined`
and trying to call `toString()`) for non-string args, string args just
have preferential treatment now.

Yes, a stringable non-string value now has worse performance because of
the extra `isString()` call at the start of `checkedToString`. This is
considered a trade-off worth making because, again, arguments passed to
this library will nearly always be strings. In fact, when using
TypeScript users can only provide args of type `string` (see for example
`index.d.ts` line 86).

The new `illegalArguments` case was added the cover the scenario where
the final if-statement in `maybeToString` is written in a way that does
not handle non-null (and non-string) return values of `toString()`.

Lastly, the new tests in `escape.test.js` and `quote.test.js` address
the case where `.toString()` of strings somehow doesn't return a string.
I think these test makes sense because I think it would be unexpected if
a string is rejected because its `.toString()` value isn't a string
(which previously it would! because of the obscurity of this issue, I'm
not considering it a bugfix nor a breaking change though).

Ow, and this re-orders imports in `main.js` to be alphabetical.