Remove `src/main.js` (#971)
Remove `main.js` and inline it's logic into `index.js`. The motivation
is threefold:
- It motivates improvements to the integration tests, as seen in this
  commit.
- It aligns the v1 code base close to the v2 code base making it easier
  to maintain them side-by-side.
- It removes an abstraction layer that mainly serves to enable safe unit
  testing through dependency injection. This is not inherently bad, but
  in this case the module did not add much more than that.

Regarding the first point (improvements to the integration tests), this
manifests in a couple of ways. First and foremost, quoting functionality
now has integration coverage for the `flagProtection` option. Secondly,
both escaping and quoting functionality now have coverage that disabling
`flagProtection` actually disables that functionality (due to new
fixtures in the respective categories).

Moreover, the integration tests are further improved by 1) removing
unnecessary macros, 2) aligning escape and quote tests, and 3) general
cleaning up of integration tests (e.g. removing duplicate tests).

Lastly, a note on the disabled ObjectLiteral mutant (when getting the
default shell). The value passed is not relevant on Unix systems, hence
in CI this mutant can never be caught. An improvement over disabling the
mutant would be welcome in the future.