Update escaping for PowerShell with `{interpolation:false}` (#984)
* Don't escape $ for PowerShell without interpolation

The escape logic for PowerShell with `interpolation: false` (not quote)
would escape "$" as "`$". This escaping was unnecessary for legitimate
use cases and would lead to the exact value "`$" being received by the
invoked program.

* Don't escape ` for PowerShell without interpolation

The escape logic for PowerShell with `interpolation: false` (not quote)
would escape "`" as "`$`. This escaping was unnecessary for legitimate
use cases and would lead to the exact value "``" being received by the
invoked program, instead of the expected "`".

* Update the changelog