Update quoting for CMD to ... not quote(?)
This commit is a work in progress...

CMD quoting is weird - It appears to be necessary to preserve whitespace
but you can't disable variable expansion from within quotes (evidenced
by the removal of `_common.cjs` line 68 to 70).

The main progress of this commit is that it fixes the incorrect escaping
of `%` when quoting for CMD. However, it lacks whitespace preservation
which should really be kept - and, to reiterate, whitespace preservation
is impossible without quoting in CMD, as far as I'm aware.