Do quote for CMD!
Turns out quoting is necessary in CMD for one important reason: to
prevent argument splitting.

The side affect of quoting whitespace seems to be that, in contrast to
escaping for interpolation, double quotes must be escaped differently.
Hence, this commit also re-introduces the `escapeArgForQuoted` function.
This particular escaping gets a bit confusing when a pre-existing quote
is next to a whitespace character because the inserted quotes from the
actual quoting have an effect on those pre-existing quotes, so they need
to be handled separately.