Move CI auditing into separate workflow (#992)
Separate the continuous audit job from the general checks workflow. The
motivation for this is that checks should always pass (similar to `npm
run verify`), but audit may not always be passing and this is not
always a problem.

One place where this has an immediate impact is the build status badge
in the README, which will no longer show as failing when the audit job
is failing.