Update quoting for CMD (#986)
Update quoting for CMD to only qoute whitespace. The trigger for this
change is the incorrect output when quoting an input string containing a
`%` (evidenced by the removal of `_common.cjs` line 68 to 70). Based on
resources explaining how CMD quoting works as well as experimentation it
seems to me best to quote only whitespace for CMD, quoting whitespace
serves two purposes:
1. It preserves the whitespace as is instead of condensing it to one
   singular space.
2. It prevents argument splitting.

Other than whitespace, argument quoting for CMD will not quote anything
and instead escape similarly to escaping for interpolation. It's not
exactly the same because the presence of quotes (for quoting purposes,
not escaped themselves) changes how quotes within the arguments have to
be escaped.

Notably a bunch of seemingly unrelated tests for quoting have been
updated to no longer assert that quoted strings start and end with a
quote the reason for this is simple: CMD proves that quoting does not
necessarily imply putting a quote at the start and end of the string.
Thus, this can no longer be asserted as a property (i.e. in a property
test).