Reduce integration tests coverage of transpiled library
Remove the duplication of all integration tests for both the ESM and CJS
version of the library in favor of testing the ESM version extensively
and relying only on the `ESM === CJS` property test (as well as the
general assumption that the transpiling software works correctly) to
verify the CJS version is correct as well.

The motivation for this is to speed up integration tests and reduce
unnecessary work.