Use flag fixtures in unit tests
Update the coverage of flag protection in unit tests to be based
primarily on the flag fixtures previously used only in integration
tests. This helps increase confidence in correctness before integration
tests are run by virtue of having exact examples that _must_ be handled
correctly.

The existing coverage of this functionality has been removed as those
tests closely mimicked the behavior of the implementation (through very
similar regular expressions).