Move Semgrep into separate workflow (#1000)
Separate the continuous Semgrep job from the general checks workflow.
The motivation for this is that semgrep requires access to a secret to
work, and this secret isn't available on Pull Requests from forks. This
leads in a failed pipeline for external contributors, which is not
desired. Hence, just run it on main and address problems it finds
separately.